requirements.txt
openai
langchain
tiktoken
unstructured
pypdf
python-dotenv
streamlit
unstructured_inference
pytesseract
tesseract